Create Electric Math
Perform the math calculation stated by the server on the numbers (that's right, numbers) given by the server. possible maths are add, subtract, multiply, or divide. multiply 8 and 4. divide 8 by 4. add 8 and 4. subtract 8 from 4. send, rounded, to 8 places to the right of the decimal (0.00000001) if there is a decimal. Send whole numbers as they appear. 1, 1.05000000, 0.10000008 and the server will respond with a flag. nc listen.runcode.ninja 9001

Connecting to this IP/Port will output something like:
ohaithar, Mr. User.
multiply 8 and 4!
To which you should send:
32