Create auth-parser
A friend of yours seems to think his Minecraft server has been getting a ton of SSH brute force attempts. He wants to start blocking IP addresses that attempt to log in to his server too much. He wants to see all valid IP addresses in that log file. He's requested that you make the output similar to a manual SQL query ran from the command line so he can review the output to know which IPs to block. Can you parse the log file and mimic the output? Pay close attention to the formatting, attention to detail is paramount.

# ./auth_solve.py /path/to/file/auth1.log.gz
-------------------------------------
| Percent | Count |              IP |
-------------------------------------
|  0.0563 |     1 |  91.121.158.124 |
|  0.0563 |     1 |    146.0.75.199 |
|  0.0563 |     1 |  163.172.69.249 |
|  0.0563 |     1 |  178.15.105.238 |
|  0.0563 |     1 |   185.156.177.8 |
|  0.0563 |     1 |  185.156.177.53 |
|  0.0563 |     1 |  185.222.211.18 |
|  0.0563 |     1 |    196.52.43.94 |
|  0.0563 |     1 | 213.202.230.144 |
|  0.1126 |     2 |    5.39.216.134 |
|  0.1126 |     2 |    71.6.165.200 |
|  0.1690 |     3 |      5.8.10.202 |
|  0.2253 |     4 |  80.211.140.131 |
|  0.2816 |     5 |   41.59.225.121 |
|  0.2816 |     5 |    42.159.246.3 |
|  0.2816 |     5 | 123.150.200.121 |
|  0.2816 |     5 |   138.19.133.51 |
|  0.2816 |     5 |    168.70.37.89 |
|  0.2816 |     5 |  197.245.39.234 |
|  0.3380 |     6 |  151.40.221.171 |
|  0.5633 |    10 |  112.112.102.38 |
|  0.5633 |    10 |  178.68.209.126 |
|  0.5633 |    10 | 221.148.234.252 |
|  1.1267 |    20 |    5.255.68.179 |
|  1.4647 |    26 | 192.129.227.186 |
|  3.9436 |    70 |     87.27.92.12 |
|  16.000 |   284 |     137.74.1.62 |
|  20.169 |   358 |  212.237.47.236 |
|  25.352 |   450 |   185.182.56.85 |
|  27.098 |   481 |     80.237.75.2 |
-------------------------------------
|   Total |  1775 |                 |
-------------------------------------